hawky/requirements.txt
discord.py
asyncpg
dotenv
